Create Main.py
This is the main file that takes in input from webcam and performs all the methods, it and identifies and labels the key points of the hand. After all the key point positions are recorded in a list, it is then transferred to unity engine through a socket plug-in at server port 5052.